python/pyproject.toml
[tool.poetry]
name = "grpc-python"
version = "1"
description = ""
authors = ["Shubham <shubhammahawar.26@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
grpcio = ">=1.71.0,<2.0.0"
grpcio-tools = ">=1.71.0,<2.0.0"
grpcio-reflection = ">=1.71.0,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
